pyproject.toml
[project]
name = "socialeval"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "redis>=5.0",
    "celery>=5.4",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "pyyaml>=6.0",
    "jsonschema>=4.23",
    "openai>=1.58",
    "anthropic>=0.42",
    "matplotlib>=3.9",
    "weasyprint>=62.3",
    "jinja2>=3.1",
    "python-jose[cryptography]>=3.3",
    "itsdangerous>=2.2",
    "passlib[bcrypt]>=1.7",
    "python-multipart>=0.0.12",
    "aiosmtplib>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
